+++ scripts/format.sh
@@ -19,8 +19,8 @@ elif [[ ${OS} = "Darwin" ]] ; then
 fi
 
 # Discover clang-format
-if type clang-format-10.0.0 2> /dev/null ; then
-    CLANG_FORMAT=clang-format-10.0.0
+if type clang-format-10 2> /dev/null ; then
+    CLANG_FORMAT=clang-format-10
 elif type clang-format 2> /dev/null ; then
     # Clang format found, but need to check version
     CLANG_FORMAT=clang-format
@@ -30,7 +30,7 @@ elif type clang-format 2> /dev/null ; then
         #exit 1
     fi
 else
-    echo "No appropriate clang-format found (expected clang-format-10.0.0, or clang-format)"
+    echo "No appropriate clang-format found (expected clang-format-10, or clang-format)"
     exit 1
 fi